Bug Fixing

+++ TradesCompany.Domain/Entities/Bill.cs
@@ -10,10 +10,10 @@ namespace TradesCompany.Domain.Entities
     {
         public int Id { get; set; }
         public string Title { get; set; } = null!;
-        public double serviceCharge { get; set; }
-        public double Gst {  get; set; }
-        public double PlatFormFees { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal serviceCharge { get; set; }
+        public decimal Gst {  get; set; }
+        public decimal PlatFormFees { get; set; }
+        public decimal TotalPrice { get; set; }
         public string? FilePath { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.Now;
         public int serviceScheduleId { get; set; }
